feat: add service layer + cleanup
Add service layer for LNGAgeing and ship
Remove unused repositories
Rename ResultRow to CalculationStep
Set unique constraint on ship name+country
