--- api/src/main/kotlin/repositories/StandardRepository.kt
@@ -1,8 +1,0 @@
-package com.equinor.neqsim.repositories
-
-import com.equinor.neqsim.entities.Standard
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface StandardRepository : JpaRepository<Standard, Long>
