--- api/src/main/kotlin/repositories/TransportRepository.kt
@@ -1,8 +1,0 @@
-package com.equinor.neqsim.repositories
-
-import com.equinor.neqsim.entities.Transport
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface TransportRepository : JpaRepository<Transport, Long>
