+++ api/src/main/kotlin/resolvers/CalculationMutationResolver.kt
@@ -2,15 +2,13 @@ package com.equinor.neqsim.resolvers
 
 import com.coxautodev.graphql.tools.GraphQLMutationResolver
 import com.equinor.neqsim.entities.*
-import com.equinor.neqsim.repositories.CalculationRepository
-import com.equinor.neqsim.tools.*
-import neqsim.fluidMechanics.flowSystem.twoPhaseFlowSystem.shipSystem.LNGship
-import neqsim.thermo.system.SystemInterface
-import neqsim.thermo.system.SystemSrkEos
+import com.equinor.neqsim.services.LNGAgeingService
+import com.equinor.neqsim.utils.FluidInput
 import org.springframework.stereotype.Component
+import java.time.OffsetDateTime
 
 @Component
-class CalculationMutationResolver(private val calculationRepository: CalculationRepository) : GraphQLMutationResolver {
+class CalculationMutationResolver(private val lngAgeingService: LNGAgeingService) : GraphQLMutationResolver {
 
     fun addCalculation(
         shipInput: ShipInput,
@@ -54,55 +52,25 @@ class CalculationMutationResolver(private val calculationRepository: Calculation
             idealGasReferenceState = standardInput.idealGasReferenceState
         )
 
-        val calculation = Calculation(
-            ship = ship,
-            fluid = fluid,
-            transport = transport,
-            standard = standard
-        )
-        ship.calculations.add(calculation)
-        fluid.calculation = calculation
-        transport.calculation = calculation
-        standard.calculation = calculation
-
-        // LNG ship calculation
-        val system: SystemInterface = SystemSrkEos(273.15 - 161.4, transport.pressure)
-        system.addComponent("nitrogen", fluid.nitrogen)
-        system.addComponent("methane", fluid.methane)
-        system.addComponent("ethane", fluid.ethane)
-        system.addComponent("propane", fluid.propane)
-        system.addComponent("iButane", fluid.iButane)
-        system.addComponent("nButane", fluid.nButane)
-        system.addComponent("iPentane", fluid.iPentane)
-        system.addComponent("nPentane", fluid.nPentane)
-        system.addComponent("nHexane", fluid.nHexane)
-        val lngShip = LNGship(system, transport.volume, transport.boilOffRate).apply {
-            standardISO6976.energyRefT = standard.combustionTemperature
-            standardISO6976.volRefT = standard.measurementTemperature
-            standardISO6976.referenceState = if (standard.idealGasReferenceState) "ideal" else "real"
-        }
-        lngShip.createSystem()
-        lngShip.init()
-        lngShip.solveSteadyState(0)
-        lngShip.solveTransient(0)
-        val result = lngShip.getResults("result")
-
-        for (resultRow in result.drop(1)) {
-            calculation.result.add(
-                ResultRow(
-                    time = stringToDouble(resultRow[1]),
-                    temp = stringToDouble(resultRow[2]),
-                    wi = stringToDouble(resultRow[3]),
-                    gcv = stringToDouble(resultRow[4]),
-                    density = stringToDouble(resultRow[5]),
-                    volume = stringToDouble(resultRow[6]),
-                    energy = stringToDouble(resultRow[7]),
-                    calculation = calculation
-                )
-            )
-        }
-
-        calculationRepository.save(calculation)
-        return calculation
+        return lngAgeingService.createCalculation(ship, fluid, transport, standard)
     }
 }
+
+data class ShipInput (
+    val name: String,
+    val country: String
+)
+
+class StandardInput (
+    val combustionTemperature: Double,
+    val measurementTemperature: Double,
+    val idealGasReferenceState: Boolean
+)
+
+class TransportInput (
+    val volume: Double,
+    val pressure: Double,
+    val boilOffRate: Double,
+    val toDate: OffsetDateTime,
+    val fromDate: OffsetDateTime
+)
