+++ api/src/main/kotlin/resolvers/ShipQueryResolver.kt
@@ -2,13 +2,13 @@ package com.equinor.neqsim.resolvers
 
 import com.coxautodev.graphql.tools.GraphQLQueryResolver
 import com.equinor.neqsim.entities.Ship
-import com.equinor.neqsim.repositories.ShipRepository
+import com.equinor.neqsim.services.ShipService
 import org.springframework.stereotype.Component
 
 @Component
-class ShipQueryResolver(private val shipRepository: ShipRepository): GraphQLQueryResolver {
+class ShipQueryResolver(private val shipService: ShipService): GraphQLQueryResolver {
 
     fun ships(): MutableIterable<Ship> {
-        return shipRepository.findAll()
+        return shipService.findAll()
     }
 }
