+++ api/src/main/kotlin/utils/StringUtils.kt
@@ -1,4 +1,4 @@
-package com.equinor.neqsim.tools
+package com.equinor.neqsim.utils
 
 fun stringToDouble(string: String): Double {
     return string.replace(",","").toDouble()
