+++ api/src/main/resources/db/migration/V2__rename_result_row_table.sql
@@ -1,0 +1,3 @@
+ALTER TABLE result_row RENAME TO calculation_step;
+CREATE UNIQUE INDEX uc_ship
+    ON ship (name, country);
